rename config mod to decode
