+++ dnp3/examples/master_tcp_client.rs
@@ -2,7 +2,7 @@ use dnp3::app::enums::OpType;
 use dnp3::app::retry::ReconnectStrategy;
 use dnp3::app::timeout::Timeout;
 use dnp3::app::variations::{Group12Var1, Variation};
-use dnp3::config::*;
+use dnp3::decode::*;
 use dnp3::link::{EndpointAddress, LinkErrorMode};
 use dnp3::master::association::AssociationConfig;
 use dnp3::master::handle::{Listener, MasterConfig, NullHandler};
