+++ dnp3/examples/outstation_tcp_server.rs
@@ -1,7 +1,7 @@
 use dnp3::app::enums::CommandStatus;
 use dnp3::app::flags::Flags;
 use dnp3::app::measurement::*;
-use dnp3::config::*;
+use dnp3::decode::*;
 use dnp3::link::{EndpointAddress, LinkErrorMode};
 use dnp3::outstation::config::OutstationConfig;
 use dnp3::outstation::database::config::*;
@@ -9,6 +9,7 @@ use dnp3::outstation::database::{Add, Update, UpdateOptions};
 use dnp3::outstation::database::{EventBufferConfig, EventClass};
 use dnp3::outstation::tcp::{AddressFilter, TcpServer};
 use dnp3::outstation::traits::*;
+
 use std::time::Duration;
 
 fn get_outstation_config() -> OutstationConfig {
