+++ dnp3/src/decode.rs
