+++ dnp3/src/link/format.rs
@@ -1,4 +1,4 @@
-use crate::config::LinkDecodeLevel;
+use crate::decode::LinkDecodeLevel;
 use crate::link::constant;
 use crate::link::crc::{calc_crc, calc_crc_with_0564};
 use crate::link::display::LinkDisplay;
