+++ dnp3/src/link/layer.rs
@@ -1,5 +1,5 @@
 use crate::app::EndpointType;
-use crate::config::DecodeLevel;
+use crate::decode::DecodeLevel;
 use crate::link::error::LinkError;
 use crate::link::function::Function;
 use crate::link::header::{
