+++ dnp3/src/outstation/config.rs
@@ -1,4 +1,4 @@
-use crate::config::DecodeLevel;
+use crate::decode::DecodeLevel;
 use crate::link::EndpointAddress;
 use crate::outstation::database::ClassZeroConfig;
 use crate::util::buffer::Buffer;
