+++ dnp3/src/outstation/tests/harness/harness.rs
@@ -1,4 +1,4 @@
-use crate::config::AppDecodeLevel;
+use crate::decode::AppDecodeLevel;
 use crate::link::header::{BroadcastConfirmMode, FrameInfo, FrameType};
 use crate::outstation::config::{Feature, OutstationConfig};
 use crate::outstation::database::EventBufferConfig;
