+++ dnp3/src/transport/reader.rs
@@ -1,5 +1,5 @@
 use crate::app::parse::parser::ParsedFragment;
-use crate::config::{AppDecodeLevel, DecodeLevel};
+use crate::decode::{AppDecodeLevel, DecodeLevel};
 use crate::link::error::LinkError;
 use crate::link::{EndpointAddress, LinkErrorMode};
 use crate::outstation::config::Feature;
