+++ dnp3/src/transport/real/reader.rs
@@ -1,5 +1,5 @@
 use crate::app::EndpointType;
-use crate::config::DecodeLevel;
+use crate::decode::DecodeLevel;
 use crate::link::error::LinkError;
 use crate::link::header::FrameType;
 use crate::link::parser::FramePayload;
